Change the Match component in the es6 example to use properties.
The big difference between the es5 and the es6 syntax of this example is
that defaultProps is defined as a property of the constructor in ES6,
while they are defined in the class body in ES5.

Notice that in order to set these properties on our component, we have
to define the constructor function that invokes super.

Note: The tutorial at
https://facebook.github.io/react/docs/reusable-components.html contains
valuable information on how to write ReactJS in ES6 vs. ES5.